export/requirements.txt
google-cloud-storage
pyyaml
duckdb
